refactor: create common folder in order to keep resources shared between modules
