+++ src/backend/src/modules/common/queries/createGamePlatform.js
