+++ src/backend/src/modules/common/queries/deleteGamePlatform.js
